requirements.txt
streamlit
pandas
langchain
langchain-google-genai
google-generativeai
python-dotenv
openpyxl
matplotlib
